requirements.txt
pdfminer.six>=20221105
pdf2image>=1.16.3
pytesseract>=0.3.10
pypandoc>=1.11
pandas>=1.5.3
openpyxl>=3.1.2
tabulate>=0.9.0